+++ paper/paper.md
@@ -9,7 +9,7 @@ tags:
 authors:
   - name: Tian Tian
     orcid: 0000-0003-0634-0455
-	corresponding: true
+    corresponding: true
     affiliation: [1,3]
   - name: Lucas R Timmerman
     orcid: 0000-0001-5664-5762
@@ -19,7 +19,7 @@ authors:
     affiliation: 1
   - name: Ben Comer
     orcid: 0000-0002-7528-0049
-	affiliation: 1
+    affiliation: 1
   - name: Andrew J Medford
     orcid: 0000-0001-8311-9581
     corresponding: true
@@ -30,13 +30,13 @@ authors:
     affiliation: [1, 2]
 affiliations:
   - name: College of Engineering, Georgia Institute of Technology, Atlanta, GA 30332, USA
-	index: 1
+    index: 1
   - name: College of Computing, Georgia Institute of Technology, Atlanta, GA 30332, USA
     index: 2
   - name: Department of Chemical and Materials Engineering, University of Alberta, Edmonton AB, T6G 2R3, Canada
     index: 3
 
-date: 18 November 2024
+date: 05 December 2024
 bibliography: paper.bib
 ---
 
