+++ paper/paper.bib
@@ -22,7 +22,8 @@
 	urldate = {2024-01-10},
 	volume = {185},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S001046551300372X},
-	url = {https://doi.org/10.1016/j.cpc.2013.10.027}
+	url = {https://doi.org/10.1016/j.cpc.2013.10.027},
+	doi = {10.1016/j.cpc.2013.10.027},
 	}
 
 @article{kapil_i-pi-2.0_2019,
@@ -38,7 +39,8 @@
 	urldate = {2024-09-19},
 	volume = {236},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S0010465518303436},
-	url = {https://doi.org/10.1016/j.cpc.2018.09.020}}
+	url = {https://doi.org/10.1016/j.cpc.2018.09.020}},
+	doi = {10.1016/j.cpc.2018.09.020},
 
 @article{litman_i-pi-3.0_2024,
 	author = {Litman, Yair and Kapil, Venkat and Feldman, Yotam M. Y. and Tisi, Davide and Begu{\v s}i{\'c}, Tomislav and Fidanyan, Karen and Fraux, Guillaume and Higer, Jacob and Kellner, Matthias and Li, Tao E. and P{\'o}s, Eszter S. and Stocco, Elia and Trenins, George and Hirshberg, Barak and Rossi, Mariana and Ceriotti, Michele},
@@ -54,7 +56,9 @@
 	urldate = {2024-09-18},
 	volume = {161},
 	bdsk-url-1 = {https://pubs.aip.org/jcp/article/161/6/062504/3308034/i-PI-3-0-A-flexible-and-efficient-framework-for},
-	url = {https://doi.org/10.1063/5.0215869}}
+	url = {https://doi.org/10.1063/5.0215869},
+	doi = {10.1063/5.0215869},
+	}
 
 @article{xu_m-sparc-1.0_2020,
 	author = {Xu, Qimen and Sharma, Abhiraj and Suryanarayana, Phanish},
@@ -68,7 +72,9 @@
 	title = {M-{SPARC}: Matlab-Simulation Package for Ab-initio Real-space Calculations},
 	volume = 11,
 	year = 2020,
-	url = {http://dx.doi.org/10.1016/j.softx.2020.100423}}
+	url = {http://dx.doi.org/10.1016/j.softx.2020.100423},
+	doi = {10.1016/j.softx.2020.100423},
+	}
 
 @article{zhang_m-sparc-2.0_2023,
 	author = {Zhang, Boqin and Jing, Xin and Kumar, Shashikant and Suryanarayana, Phanish},
@@ -81,7 +87,9 @@
 	title = {Version 2.0.0 - M-SPARC: Matlab-Simulation Package for Ab-initio Real-space Calculations},
 	volume = 21,
 	year = 2023,
-	url = {http://dx.doi.org/10.1016/j.softx.2022.101295}}
+	url = {http://dx.doi.org/10.1016/j.softx.2022.101295},
+	doi = {10.1016/j.softx.2022.101295},
+	}
 
 @article{xu_sparc-1.0_2021,
 	author = {Xu, Qimen and Sharma, Abhiraj and Comer, Benjamin and Huang, Hua and Chow, Edmond and Medford, Andrew J. and Pask, John E. and Suryanarayana, Phanish},
@@ -94,7 +102,9 @@
 	title = {SPARC: Simulation Package for Ab-initio Real-space Calculations},
 	volume = 15,
 	year = 2021,
-	url = {http://dx.doi.org/10.1016/j.softx.2021.100709}}
+	url = {http://dx.doi.org/10.1016/j.softx.2021.100709},
[…]
