+++ paper/paper.bib
@@ -39,8 +39,9 @@
 	urldate = {2024-09-19},
 	volume = {236},
 	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S0010465518303436},
-	url = {https://doi.org/10.1016/j.cpc.2018.09.020}},
+	url = {https://doi.org/10.1016/j.cpc.2018.09.020},
 	doi = {10.1016/j.cpc.2018.09.020},
+	}
 
 @article{litman_i-pi-3.0_2024,
 	author = {Litman, Yair and Kapil, Venkat and Feldman, Yotam M. Y. and Tisi, Davide and Begu{\v s}i{\'c}, Tomislav and Fidanyan, Karen and Fraux, Guillaume and Higer, Jacob and Kellner, Matthias and Li, Tao E. and P{\'o}s, Eszter S. and Stocco, Elia and Trenins, George and Hirshberg, Barak and Rossi, Mariana and Ceriotti, Michele},
