### 2025-01-16 #7565 @RayLei-TRI
Hi Bonan,

  Thank you so much for the invitation! I would very much love to
contribute to JOSS, but for this particular group and project I think there
is a conflict of interest, as I am a former member of that team.

  I would love to help with other reviews down the line though.

Best,
Ray
