### 2025-01-17 #7565 @zhubonan
> Hi Bonan,
>
>   Thank you so much for the invitation! I would very much love to
> contribute to JOSS, but for this particular group and project I think there
> is a conflict of interest, as I am a former member of that team.
>
>   I would love to help with other reviews down the line though.
>
> Best,
> Ray

Thanks for letting me know and disclosing the conflict of interest. 👏
It is possible to waive the conflict if you think you make impartial judgement of this work.
The policy and some examples can be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy
