### 2025-02-05 #7747 @editorialbot
[REVIEW]: SPARC-X-API: Versatile Python Interface for Real-space Density Functional Theory Calculations
**Submitting author:** @alchem0x2a (<a href="http://orcid.org/0000-0003-0634-0455">Tian Tian</a>)
**Repository:** https://github.com/SPARC-X/SPARC-X-API
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.5
