### 2025-02-23 SPARC-X/SPARC-X-API#89 @naik-aakash
Documentation enhancement suggestions
Dear developers,

I have started the review for the Joss submission https://github.com/openjournals/joss-reviews/issues/7747

It is already very detailed and comprehensive. As I was walking through the documentation, I noticed the following things, and I think they might help to make it a bit clearer.

1. The command below did not work for me.
https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/doc/installation.md?plain=1#L20
 I am using miniconda to create and manage my conda environments. Had  to specify `conda-forge` channel to get it working. Maybe better to specifically add this channel in the command?

2. Could it be better to add a note to already alert the users that pre-compiled binaries from conda does not have socket interface enabled. I see it is mentioned later in advanced section.

3. Noticed some **TODO** section comments rendered in the deployed docs, seems it was supposed to be for internal reference, but somehow accidentally got added in final documentation.

4. I came across minor typos in the following lines in [paper.md](https://github.com/SPARC-X/SPARC-X-API/blob/master/paper/paper.md). I'm listing here some general ones I noticed. There are additional places where articles are missing or commas are added in text where they may not be needed.
[protol](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L69C18-L69C29)
[unargubaly](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L79)
[scalable,](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L101) -> comma not needed?
[both](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L103) -> can be remove from the sentence there?
[comatible](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L160)
[is](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L240) -> accidental is in the sentence, needs to be removed
