### 2025-02-23 SPARC-X/SPARC-X-API#90 @naik-aakash
Open Questions
Part of https://github.com/openjournals/joss-reviews/issues/7747

 1. Is there any specific reason to use `setup.py` instead of `project.toml`?
 2. Can optional dependencies of specific versions supported or tested with sparc using plumed, mace , finetuna added ?
 3. I could not figure out how to use train for example MACE or any of models from fairchem with SPARC, can an small example added to demonstrate it?
