### 2025-03-01 SPARC-X/SPARC-X-API#91 @vijaymocherla
Issues with sparc doc-parser
sparc `doc-parser` generates too may user warnings when configured with `ase`. I've attached the errors as file.

[sparc_import_error.txt](https://github.com/user-attachments/files/19038581/sparc_import_error.txt)

Following is the `sparc` configuration in `config.ini`:
```
[sparc]
; `command`: full shell command (include MPI directives) to run SPARC
command = ~/miniconda3/envs/sparc/bin/sparc

; `psp_path`: directory containing pseudopotential files (optional)
psp_path = ~/sparc-x-api-review/SPMS-psps

; `doc_path`: directory for SPARC LaTeX documentation to build JSON schema on the fly (optional)
doc_path = ~/miniconda3/envs/sparc/doc/.LaTeX/

```
