### 2025-03-01 SPARC-X/SPARC-X-API#92 @vijaymocherla
Errors running the socket example
I've encountered the following error, while trying to work my way through the examples.  This specific error is from running `./examples/simple_examples/ex1-ase-optimize.py`, for a complete output please see the attached output file.

```
  File "~/miniconda3/envs/sparc/lib/python3.13/site-packages/sparc/calculator.py", line 628, in _calculate_with_socket
    raise RuntimeError(
        "Your sparc binary is not compiled with socket support!"
    )
RuntimeError: Your sparc binary is not compiled with socket support!
```

[ex1_output.txt](https://github.com/user-attachments/files/19038595/ex1_output.txt)

I've encountered a similar error while trying to test socket connection based examples with `ase`. I have installed both `sparc` and `sparc-x-api` from the `conda-forge` channel. However, I can confirm that I could find the sparc binary from CLI and run `ex0-eos.py`.
