### 2025-03-01 SPARC-X/SPARC-X-API#92 @vijaymocherla
Additionally, could you please comment on your decision to make the ase calculator for `sparc` as a standalone package rather than being native implemented within `ase` like the calculators that exist for various other packages.
