### 2025-03-04 SPARC-X/SPARC-X-API#93 @jacksund
conda installation fails to find mamba on default channel
The first command for installation fails with the following:
```
(base) jacksund@jacksund-ThinkPad-E14-Gen-3:~$ conda create -n sparc-env python=3.11 pip mamba
Retrieving notices: ...working... done
Channels:
 - defaults
Platform: linux-64
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - mamba

Current channels:

  - defaults

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.
```

Adding `-c conda-forge` to the command fixes it. I'm guessing you have this in your conda defaults but just forgot to add it to the docs
