### 2025-03-04 SPARC-X/SPARC-X-API#94 @jacksund
are pip and mamba required?
the first install command given is `conda create -n sparc-env python=3.11 pip mamba`, but are pip and mamba actually needed? I believe pip is installed alongside python, and libmamba is now the default solver for conda.

The install section could probably be reduced to:

```
conda create -n sparc-env -c conda-forge python=3.11 sparc-x-api
```
