### 2025-03-04 SPARC-X/SPARC-X-API#94 @jacksund
if they are actual dependencies, maybe you can them to your conda feedstock...?
