### 2025-03-04 SPARC-X/SPARC-X-API#95 @jacksund
several TODO mentions in docs
> TODO MKL available?

> TODO make sure modules are correct.

> TODO add module in script.

^^^ examples on the installation page. Should probably hide these throughout the docs 😄
