### 2025-03-04 SPARC-X/SPARC-X-API#96 @jacksund
many warnings listed on first import
```
>>> import sparc
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:133: UserWarning: TeX file /home/jacksund/anaconda3/envs/sparc-env/doc/sparc/.LaTeX/SQ.tex is missing! It may be a typo in the document, ignore parameters from this file.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:159: UserWarning: Cannot find the c source file "initialization.c", skip version parsing!
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:204: UserWarning: Provided a non-structured frame for parsing, skip.
  warn("Provided a non-structured frame for parsing, skip.")
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in 0.0 for odd k-point mesh
 0.5 for even k-point mesh cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in 2320.904 for gaussian
1160.452 for fermi-dirac cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in 0.007350 for gaussian
0.003675 for fermi-dirac cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in N_e/2 x 1.2 + 5 cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in see description cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in h^2x0.001 cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:220: UserWarning: Parameter PRECOND_RESTA_Q0 is disabled in the doc, ignore!
  warn(f"Parameter {symbol} is disabled in the doc, ignore!")
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:220: UserWarning: Parameter PRECOND_RESTA_RS is disabled in the doc, ignore!
  warn(f"Parameter {symbol} is disabled in the doc, ignore!")
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:220: UserWarning: Parameter PRECOND_FITPOW is disabled in the doc, ignore!
  warn(f"Parameter {symbol} is disabled in the doc, ignore!")
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in 0.2*TOL_SCF cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in \max(TOL_FOCKx 10,0.001) cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in 0.25 for PBE0 and HSE cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in TOL_SCFx 0.01 cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in TOL_SCFx 0.001 cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in ION_TEMP cannot converted to a numerical array, will skip conversion.
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:220: UserWarning: Parameter ION_ELEC_EQT is disabled in the doc, ignore!
  warn(f"Parameter {symbol} is disabled in the doc, ignore!")
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:727: UserWarning: Type of NPT_NH_QMASS if not standard digit or array, mark as others.
  warn(f"Type of {symbol} if not standard digit or array, mark as others.")
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/docparser.py:573: UserWarning: Some fields in Generated by a linear model. cannot converted to a numerical array, will skip conversion.
  warn(
```

some of these warnings make me think the issue is with the pre-compiled sparc-x install, where I used `conda install -c conda-forge sparc-x` from the docs.

I'm going to revisit the conda toolchains section of the installation and see if that works
