### 2025-03-04 SPARC-X/SPARC-X-API#97 @jacksund
`Bulk` is now `bulk` in ase v3.24
following along with the basic example here:
https://sparc-x.github.io/SPARC-X-API/basic_usage.html

```
>>> from ase.build import Bulk
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ImportError: cannot import name 'Bulk' from 'ase.build' (/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/ase/build/__init__.py)
```

Looks like the most recent ase version uses `bulk`, not `Bulk`: https://gitlab.com/ase/ase/-/blob/master/ase/build/bulk.py?ref_type=heads#L17

I have ase v3.24.0 installed for reference. If this isn't correct, then this line in the conda feedstock should be updated: https://github.com/conda-forge/sparc-x-api-feedstock/blob/main/recipe/meta.yaml#L29
