### 2025-03-04 SPARC-X/SPARC-X-API#98 @jacksund
directory kwarg errors and extra warnings
looks like `directory="run_sp"` errors because it's not a known kwarg -- maybe this is another ASE version issue? I'll switch to ase v3.22 for the rest of my testing just in case, but it still looks odd. The calculator also returns energy + forces successfully, but I still have something going on with my underlying SPARC setup. If you've seen these errors before I'd appreciate a pointer 😄 . Or if these warnings are normal, it'd be nice to silence them / turn down the logs verbosity

```
>>> from sparc.calculator import SPARC
>>> from ase.build import molecule
>>> atoms = molecule("H2", cell=(10,10,10), pbc=True, directory="run_sp")
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/ase/build/molecule.py", line 70, in molecule
    mol = Atoms(mol, **kwargs)
          ^^^^^^^^^^^^^^^^^^^^
TypeError: Atoms.__init__() got an unexpected keyword argument 'directory'
>>> atoms = molecule("H2", cell=(10,10,10), pbc=True)
>>> atoms.calc = SPARC(h=0.25)
>>> atoms.get_potential_energy()
sparc not initialized
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/calculator.py:508: UserWarning: Your sparc command is inferred to be /home/jacksund/vasp/05_openmpi/install/bin/mpirun -n 1 /home/jacksund/anaconda3/envs/sparc-env/bin/sparc, If this is not correct, please manually set $ASE_SPARC_COMMAND or SPARC.command!
  warn(
[jacksund-ThinkPad-E14-Gen-3:10540] mca_base_component_repository_open: unable to open mca_btl_openib: librdmacm.so.1: cannot open shared object file: No such file or directory (ignored)
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/sparc_parsers/utils.py:73: UserWarning: Key POISSON_SOLVER not in validator's parameter list, ignore value conversion!
  warn(
/home/jacksund/anaconda3/envs/sparc-env/lib/python3.11/site-packages/sparc/sparc_parsers/utils.py:73: UserWarning: Key VERBOSITY not in validator's parameter list, ignore value conversion!
  warn(
np.float64(-31.788191747613503)
>>> atoms.get_forces()
/home/jacksund/vasp/05_openmpi/install/bin/mpirun -n 1 /home/jacksund/anaconda3/envs/sparc-env/bin/sparc
[jacksund-ThinkPad-E14-Gen-3:10562] mca_base_component_repository_open: unable to open mca_btl_openib: librdmacm.so.1: cannot open shared object file: No such file or directory (ignored)
array([[ 4.98352144e-06,  3.40593917e-07, -1.23134459e+01],
       [-4.98352144e-06, -3.40593917e-07,  1.23134459e+01]])
```
