### 2025-03-04 SPARC-X/SPARC-X-API#89 @jacksund
I should have read this before starting my own issues 😆
number 1 -->  #93
number 3 --> #95
