### 2025-03-04 SPARC-X/SPARC-X-API#99 @jacksund
example output missing
for the simple dft example: https://sparc-x.github.io/SPARC-X-API/examples/simple_dft.html

> The output from the above example may look like:

Script runs successfully, but the output is just missing in the docs
