### 2025-03-04 SPARC-X/SPARC-X-API#100 @jacksund
make example scripts consistent
in the examples section: https://sparc-x.github.io/SPARC-X-API/examples.html

the 1st example uses a `if __name__ == "__main__"` approach to calling the script.

the 2nd example defines the main function but relies on the user to call it somewhere

the 3rd example runs the script in-line

the 4th example runs the script in-line, but the script is broken up to all for explaining each section

It'd be nice if these were in the same format 😄 . This is just a preference thing, so feel free to ignore / close this issue as well
