### 2025-03-04 SPARC-X/SPARC-X-API#101 @jacksund
replace "troubleshooting/known issues" section with FAQs
the troubleshooting section: https://sparc-x.github.io/SPARC-X-API/troubleshooting.html

I would just leave github to handle the issue tracking and now duplicate you work here. If you want to keep a section for troublehshooting, I would instead make a Frequently Asked Questions section with common errors / fixes users have
