### 2025-03-04 SPARC-X/SPARC-X-API#102 @jacksund
paper feedback
It's a nice paper!

The only comment I had while reading through was on the `Statement of Need` section -- the first paragraph and a half of this section focuses heavily on the need for SPARC-X, not the SPARC-X-API. Since the paper is about the API, I would focus more on why the SPARC-X project has a need for the python API specifically, rather than why the SPARC-X project itself exists.

Again, feel free to disregard and close this issue. Just preference + suggestion
