### 2025-03-04 SPARC-X/SPARC-X-API#103 @jacksund
why monkey patch ASE instead of contribute?
skimming through the code, it looks like you register the sparc caculator by monkey-patching ASE features:
https://github.com/SPARC-X/SPARC-X-API/blob/master/sparc/io.py#L1028

the command-line interface is also just a wrapper around the patched ase cli: https://github.com/SPARC-X/SPARC-X-API/blob/master/sparc/api.py

It's clear in the docs and paper that sparc-x-api is  *built on top of* ase, but it's not immediately obvious from the docs that running `import sparc` is actively patching ase modules. Could you add a section that explains what is done? For example, is there a chance that any ASE features break or change because of the patching? I'm sure everything is fine (it looks like you just append modules), but still it would be helpful!

Ideally, the patching goes away and the key parts are contributed to ASE instead. Is this on the roadmap? Or are some of your extra features like parameter validation holding it back?
