### 2025-03-04 #7747 @jacksund
- https://github.com/SPARC-X/SPARC-X-API/issues/93
- https://github.com/SPARC-X/SPARC-X-API/issues/94
- https://github.com/SPARC-X/SPARC-X-API/issues/95
- https://github.com/SPARC-X/SPARC-X-API/issues/96
- https://github.com/SPARC-X/SPARC-X-API/issues/97
- https://github.com/SPARC-X/SPARC-X-API/issues/98
- https://github.com/SPARC-X/SPARC-X-API/issues/99
- https://github.com/SPARC-X/SPARC-X-API/issues/100
- https://github.com/SPARC-X/SPARC-X-API/issues/101
- https://github.com/SPARC-X/SPARC-X-API/issues/102
- https://github.com/SPARC-X/SPARC-X-API/issues/103

looks like a longer list, but each issue is very small. Several issues also overlap with comments from the others too.

nice work @alchem0x2A 🥳
