### 2025-03-24 #7747 @zhubonan
Thanks @jacksund and @vijaymocherla for the detailed reviews!  @alchem0x2A can you please address the comments (issues)?
