### 2025-04-02 8467b5b
Add suppress_warnings option to SparcDocParser class
