+++ sparc/docparser.py
@@ -41,6 +41,7 @@ class SparcDocParser(object):
         parameter_categories (list): Categories of parameters extracted.
         parameters (dict): Extracted parameters with detailed information.
         other_parameters (dict): Additional parameters not categorized.
+        suppress_warnings (bool): Whether the doc parser suppress UserWarning (may be annoying during class import)
 
     Methods:
         find_main_file(main_file_pattern): Finds the main LaTeX file based on a pattern.
@@ -61,6 +62,7 @@ class SparcDocParser(object):
         intro_file="Introduction.tex",
         params_from_intro=True,
         parse_version=True,
+        suppress_warnings=True,
     ):
         """Create the doc parser pointing to the root of the doc file of SPARC
 
@@ -79,6 +81,7 @@ class SparcDocParser(object):
             intro_file: LaTeX file for the introduction
             params_from_intro: only contain the parameters that can be parsed in `intro_file`
             parse_date: get the SPARC version by date
+            suppress_warnings: whether to silence any warnings generated by the parser
         """
         self.root = Path(directory)
         self.main_file = self.find_main_file(main_file)
@@ -87,6 +90,7 @@ class SparcDocParser(object):
             raise FileNotFoundError(f"Introduction file {intro_file} is missing!")
         self.include_files = self.get_include_files()
         self.params_from_intro = params_from_intro
+        self.suppress_warnings = suppress_warnings
         self.parse_version(parse_version)
         self.parse_parameters()
         self.postprocess()
@@ -130,12 +134,13 @@ class SparcDocParser(object):
             if tex_file.is_file():
                 include_files.append(tex_file)
             else:
-                warn(
-                    (
-                        f"TeX file {tex_file} is missing! It may be a typo in the document, "
-                        "ignore parameters from this file."
+                if not self.suppress_warnings:
+                    warn(
+                        (
+                            f"TeX file {tex_file} is missing! It may be a typo in the document, "
+                            "ignore parameters from this file."
+                        )
                     )
-                )
         return include_files
 
     def parse_version(self, parse=True):
@@ -156,18 +161,20 @@ class SparcDocParser(object):
             return
         init_c = self.root.parents[1] / "src" / "initialization.c"
         if not init_c.is_file():
-            warn(
-                'Cannot find the c source file "initialization.c", skip version parsing!'
-            )
+            if not self.suppress_warnings:
+                warn(
+                    'Cannot find the c source file "initialization.c", skip version parsing!'
+                )
             self.version = None
             return
         text = open(init_c, "r", encoding="utf8").read()
         pattern_version = r"SPARC\s+\(\s*?version(.*?)\)"
         match = re.findall(pattern_version, text)
         if len(match) != 1:
-            warn(
-                'Parsing c source file "initialization.c" for version is unsuccessful!'
-            )
+            if not self.suppress_warnings:
+                warn(
+                    'Parsing c source file "initialization.c" for version is unsuccessful!'
+                )
             self.version = None
             return
         # We need to add more spacing matching in case the source code includes extra
@@ -201,7 +208,8 @@ class SparcDocParser(object):
         pattern_block = r"\\begin\{block\}\{(.*?)\}([\s\S]*?)\\end\{block\}"
         match_label = re.findall(pattern_label, frame, re.DOTALL | re.MULTILINE)
         if len(match_label) != 1:
-            warn("Provided a non-structured frame for parsing, skip.")
+            if not self.suppress_warnings:
+                warn("Provided a non-structured frame for parsing, skip.")
             return {}
         symbol, label = (
             convert_tex_parameter(match_label[0][0].strip()),
[…]
