### 2025-04-03 4272099
add test for suppressing warnings
