+++ sparc/docparser.py
@@ -83,14 +83,14 @@ class SparcDocParser(object):
             parse_date: get the SPARC version by date
             suppress_warnings: whether to silence any warnings generated by the parser
         """
+        self.suppress_warnings = suppress_warnings
+        self.params_from_intro = params_from_intro
         self.root = Path(directory)
-        self.main_file = self.find_main_file(main_file)
         self.intro_file = self.root / intro_file
+        self.main_file = self.find_main_file(main_file)
         if not self.intro_file.is_file():
             raise FileNotFoundError(f"Introduction file {intro_file} is missing!")
         self.include_files = self.get_include_files()
-        self.params_from_intro = params_from_intro
-        self.suppress_warnings = suppress_warnings
         self.parse_version(parse_version)
         self.parse_parameters()
         self.postprocess()
