### 2025-04-07 b8a7612
clearer SPARC binary options
