### 2025-04-07 452ae9a
mini section for example conda installation
