### 2025-04-07 7b1a7ae
Minimal pyproject.toml to suppress deprecation errors
