### 2025-04-07 c47540b
additional notes for conda / pip
