### 2025-04-07 99133a4
add sections about sparc installation
