### 2025-04-07 506124d
Add test output for compilation
