+++ doc/installation.md
@@ -126,8 +126,8 @@ with OpenMPI/OpenBLAS/Scalapack toolchains.
 conda activate sparc-env
 conda install -c conda-forge \
                  make compilers \
-				 fftw=*=mpi_openmpi_* \
-				 openblas openmpi scalapack
+		 fftw=*=mpi_openmpi_* \
+		 openblas openmpi scalapack
 git clone https://github.com/SPARC-X/SPARC.git
 cd SPARC/src
 # Always inspect the contents of the makefile before continue
@@ -143,6 +143,44 @@ Run a simple command to make sure the SPARC compilation works:
 ```bash
 mpirun -n 1 SPARC/lib/sparc
 ```
+
+You should see the usage info printed like following:
+```
+USAGE:
+    mpirun -np <nproc> {SPARCROOT}/lib/sparc -name <filename>
+
+    {SPARCROOT} is the location of the SPARC folder
+
+REQUIRED ARGUMENT:
+    -name <filename>
+           The filename shared by .inpt file and .ion
+           file (without extension)
+
+OPTIONS:
+    -h, --help
+           Display help (from command line).
+    -n <number of Nodes>
+    -c <number of CPUs per node>
+    -a <number of Accelerators (e.g., GPUs) per node>
+    -socket <socket>
+            <socket> can be either  <host>:<port> or <unix_socket>:UNIX.
+            Note: socket (driver) mode is an experimental feature.
+
+EXAMPLE:
+
+    mpirun -np 8 {SPARCROOT}/lib/sparc -name test
+
+    The example command runs sparc with 8 cores, with input file named
+    test.inpt, and ion file named test.ion.
+
+NOTE:
+    This is a short description of the usage of SPARC. For a detailed
+    discription, refer to the manual online at
+
+        https://github.com/SPARC-X/SPARC/tree/master/doc
+```
+
+Note that `-socket` is now an accepted option for the SPARC binary.
 
 ### Compiling SPARC on HPC
 
@@ -162,7 +200,9 @@ make USE_MKL=1 USE_SCALAPACK=0 USE_FFTW=1 USE_SOCKET=1
 ```
 
 The compiled binary will be at `SPARC/lib/sparc`, and running it
-requires the dependent modules to be loaded at runtime.
+requires the dependent modules to be loaded at runtime. You can
+perform the same quick test in the previous section to confirm the
+compilation is successful.
 
 
 Now head to the [setup tutorial](setup_environment.md) to finetune
