### 2025-04-07 4ea3d67
main function in ex1
