+++ examples/simple_examples/ex1-ase-optimize.py
@@ -79,6 +79,7 @@ def optimize_ase_bfgs_socket():
 
 
 if __name__ == "__main__":
-    # optimize_sparc_internal()
-    # optimize_ase_bfgs()
+    # Comment out specific ones if the examples run too long
+    optimize_sparc_internal()
+    optimize_ase_bfgs()
     optimize_ase_bfgs_socket()
