### 2025-04-08 2fb8391
add main function sections in the examples
