### 2025-04-08 5667bb9
update examples with downloadable script link
