+++ _sources/api/sparc.rst.txt
@@ -3,8 +3,8 @@ sparc package
 
 .. automodule:: sparc
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
 
 Subpackages
 -----------
