+++ _sources/api/sparc.client.rst.txt
@@ -3,5 +3,5 @@ sparc.client module
 
 .. automodule:: sparc.client
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
