+++ _sources/api/sparc.sparc_parsers.rst.txt
@@ -3,8 +3,8 @@ sparc.sparc\_parsers package
 
 .. automodule:: sparc.sparc_parsers
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
 
 Submodules
 ----------
