+++ _sources/api/sparc.quicktest.rst.txt
@@ -3,5 +3,5 @@ sparc.quicktest module
 
 .. automodule:: sparc.quicktest
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
