### 2025-04-08 8996161
Fix bug parsing 1d stress in MD / Geopt output
