### 2025-04-08 SPARC-X/SPARC-X-API#89 @alchem0x2A
> 3. Noticed some TODO section comments rendered in the deployed docs, seems it was supposed to be for internal reference, but somehow accidentally got added in final documentation.

The residual TODOs have been removed from the documentation #105
