### 2025-04-08 SPARC-X/SPARC-X-API#89 @alchem0x2A
> 4. I came across minor typos in the following lines in [paper.md](https://github.com/SPARC-X/SPARC-X-API/blob/master/paper/paper.md). I'm listing here some general ones I noticed. There are additional places where articles are missing or commas are added in text where they may not be needed.
[protol](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L69C18-L69C29)
[unargubaly](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L79)
[scalable,](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L101) -> comma not needed?
[both](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L103) -> can be remove from the sentence there?
[comatible](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L160)
[is](https://github.com/SPARC-X/SPARC-X-API/blob/1708de36de5e902ebba05ba64393a2f8b86be5ae/paper/paper.md?plain=1#L240) -> accidental is in the sentence, needs to be removed

I appreciate your careful examination. The corresponding typos have been corrected in #105.
