### 2025-04-08 SPARC-X/SPARC-X-API#90 @alchem0x2A
> 1. Is there any specific reason to use setup.py instead of project.toml?

Great question. The SPARC-X-API project starts a few years ago when `setup.py` was the mainstream packaging option. We would like to keep `setup.py` not only for legacy reasons but also there are some pythonic expressions which we think are better presented in `setup.py` than `pyproject.toml`
https://github.com/SPARC-X/SPARC-X-API/blob/17ea8ba552690e0020d8c762a644a87651ce538f/setup.py#L52

We're aware that while `setup.py` [will still be a valid configuration](https://packaging.python.org/en/latest/discussions/setup-py-deprecated/#), `pyproject.toml` will be the preferred solution moving forward. For now we have added a minimal `pyproject.toml` for compatibility:
https://github.com/SPARC-X/SPARC-X-API/blob/17ea8ba552690e0020d8c762a644a87651ce538f/pyproject.toml#L1-L9

We will make the transition to pyproject.toml on the roadmap for the next major release of SPARC-X-API
