### 2025-04-08 SPARC-X/SPARC-X-API#90 @alchem0x2A
>2. Can optional dependencies of specific versions supported or tested with sparc using plumed, mace , finetuna added ?
We prefer not to include optional dependencies for these external codes, as the combinations can get really complex. From our experience, if the user choose to combine SPARC-X-API with packages that rely heavily on other numerical libaries and / or pytorch, it is better to configure a workable conda environment including these packages first, and install SPARC / SPARC-X-API later since they are relatively light-weight

Please correct if I have misunderstood your question
