### 2025-04-08 SPARC-X/SPARC-X-API#90 @alchem0x2A
> 3. I could not figure out how to use train for example MACE or any of models from fairchem with SPARC, can a small example added to demonstrate it?

Training ML models like MACE / fairchem using SPARC will have the same pipeline as the official examples in these packages as SPARC-X-API is a drop-in replacement for the demo ASE calculators in these tutorials (usually `EMT`, `xtb` or `Vasp`). We didn't provide a detailed example as the main goal in this submission is to demonstrate the high-level overview of the API, and introduction of its socket interface as a convenient ASE calculator. We're happy to adapt if you think the such a demo is necessary.
