### 2025-04-08 SPARC-X/SPARC-X-API#91 @alchem0x2A
@vijaymocherla Thank you very much for your careful examination. Apologies for the long delay as I started a new position recently.

> sparc doc-parser generates too may user warnings when configured with ase. I've attached the errors as file.

You're correct. The warning information from the doc-parser is overwhelming for an end-user. #104 added a switch `suppress_warnings` for `SparcDocParser` so that the default behavior is to suppress these warnings
https://github.com/SPARC-X/SPARC-X-API/blob/17ea8ba552690e0020d8c762a644a87651ce538f/sparc/docparser.py#L65

At the end of the day, these warnings should be resolved since many of them are related to format / value conversion from the JSON schema, which we will be working in the next releases. In the meantime, suppressing these warnings should be harmless from an end-user's perspective.

During the revision we found that additional warnings could be thrown when parsing a standard MD output file, which was fixed in #108
