### 2025-04-08 SPARC-X/SPARC-X-API#92 @alchem0x2A
@vijaymocherla  Thank you for your comments.
> I've encountered the following error, while trying to work my way through the examples. This specific error is from running ./examples/simple_examples/ex1-ase-optimize.py, for a complete output please see the attached output file.

A detailed explanation can be found for the reply to #89 point 2
https://github.com/SPARC-X/SPARC-X-API/issues/89#issuecomment-2785301789

In brief, the official conda-forge package `sparc-x` only fetches the latest release in the SPARC main repo, which comes before the inclusion of the socket support. In the meantime, we have included warnings for users about the incompatibility and added sections in the documentation about how to manually compile SPARC C/C++ binary with socket support. I hope that can answer your concerns.
