### 2025-04-08 SPARC-X/SPARC-X-API#96 @alchem0x2A
@jacksund Thanks for the comment, it is also related to https://github.com/SPARC-X/SPARC-X-API/issues/91#issue-2888817465. I hope the PR #104 solves your concerns by silencing UserWarnings for an end-user. We'll still work on to make sure these warnings are fully resolved during documentation conversion, but they should be harmless for the moment
