### 2025-04-08 90da3b8
typo in readme. Bulk --> bulk
