### 2025-04-09 1be2028
fix bug that raw output disappears
