+++ _sources/examples/simple_dft.md.txt
@@ -63,7 +63,7 @@ if __name__ == "__main__":
 ```
 
 The output from the above example may look like:
-```{raw}
+```
 Fitted volume (Ang^3), energy (eV), modulus (eV/Ang^3)
 65.97840834969949 -253.07755156337953 2.9095110471623173
 Optimal cell length (cubic): 4.040799280428726 Ang
