### 2025-04-09 SPARC-X/SPARC-X-API#103 @alchem0x2A
@jacksund Yes monkeypatching was indeed the approach we add support for the sparc ioformat for ASE<=3.22, but the new external io format (EIF) introduced in ASE 3.23 made it possible to have SPARC-X-API as a standalone package which does not require messy patching of the import module. The current way of registering SPARC formats as entrypoints
https://github.com/SPARC-X/SPARC-X-API/blob/410cb1977da3719c832dedf85408bf080d5bfa96/sparc/io.py#L1162-L1191
https://github.com/SPARC-X/SPARC-X-API/blob/410cb1977da3719c832dedf85408bf080d5bfa96/setup.py#L39-L47

Since the dependencies in setup.py explicitly states `ase>=3.23`, users will always be using the new EIF approach. The monkey patch code is still there only for legacy debug purposes, and won't be invoked unless the user delibrately call `sparc.io.register_ase_io_sparc`.

We plan to remove these legacy codes in the next major release of SPARC-X-API. For now deprecation warnings are added to these functions in #109
