### 2025-04-09 SPARC-X/SPARC-X-API#103 @alchem0x2A
Regarding the contribution to ASE, we explained our reason for maintaining SPARC-X-API as an independent package in replies to issue #92, I hope that makes sense
https://github.com/SPARC-X/SPARC-X-API/issues/92#issuecomment-2785403803
