### 2025-04-09 SPARC-X/SPARC-X-API#94 @alchem0x2A
Thanks @jacksund. Your approach is the correct one. #105 adapted the corrections to the conda-forge channel. Please also find a detailed reply here
https://github.com/SPARC-X/SPARC-X-API/issues/89#issuecomment-2785285875
