### 2025-04-09 SPARC-X/SPARC-X-API#97 @alchem0x2A
This is indeed a typo, not changes in ASE. #109 fixed this, thanks for pointing out!

https://github.com/SPARC-X/SPARC-X-API/blob/410cb1977da3719c832dedf85408bf080d5bfa96/README.md?plain=1#L95-L100
