### 2025-04-09 SPARC-X/SPARC-X-API#98 @alchem0x2A
Like #97 , this is also a typo, since `directory` should only be a kwarg for ASE calculator, not Atoms object. #109 fixes this
https://github.com/SPARC-X/SPARC-X-API/blob/410cb1977da3719c832dedf85408bf080d5bfa96/README.md?plain=1#L148-L155

The extra warnings are caused by the underlying MPI library due to how it's compiled / linked https://github.com/open-mpi/ompi/issues/5280, which seems common in conda-distributed MPI. Unfortunately that's something beyond the scope of this project, but in general they should be harmless if the user choose to redirect stderr to another location. I tend to not write specific guidance about these warnings since the combination of MPI in HPC environment can be very complex, and I hope that makes sense
