### 2025-04-09 SPARC-X/SPARC-X-API#101 @alchem0x2A
Indeed. The troubleshooting section has been removed from the documentation in the latest version.
https://github.com/SPARC-X/SPARC-X-API/blob/410cb1977da3719c832dedf85408bf080d5bfa96/doc/index.md?plain=1#L9-L22
