### 2025-04-09 SPARC-X/SPARC-X-API#100 @alchem0x2A
Thanks @jacksund for the suggestions!

The way that our examples in documentation pages is to provide a breakdown of the runnable scripts in the main repo. The user needs to run a single python script to get the output results. To ensure clarity we made several changes in #105 #107 #109   :
- Make sure the full scripts can be downloaded from the documentation page via a link, like

- All examples now have a main function section for the principal logic part, like
   https://github.com/SPARC-X/SPARC-X-API/blob/410cb1977da3719c832dedf85408bf080d5bfa96/examples/FileIO/relax/relax_coords/mlff/run.py#L28-L41
