### 2025-04-09 SPARC-X/SPARC-X-API#102 @alchem0x2A
@jacksund Thanks for the suggestion. #110 slightly reduces the length of the introduction part in the statement of needs, and adds several senteces to address the needs for a API:
https://github.com/SPARC-X/SPARC-X-API/blob/7f6ed4847a295f9947405632e16f06cb5caf7f7a/paper/paper.md?plain=1#L141-L147

I hope the current modifications make sense to you. Feel free to point out if you think more changes are needed
