### 2025-04-16 f9a460c
Implementing sendposdata to avoid transpose of cell when sending data to SPARC via the socket
