### 2025-05-16 #7747 @zhubonan
Hi @alchem0x2A I can see that two out of three reviewers have completed their checklist although there are still many open issues. I think most of them can be closed. Please feel free to close them when appropriate.

Please go through the author part of the check list above. The main thing is to make a release for the version to be accepted by JOSS. Once you have the release, archive it using a repository service such as Zenodo. Please make sure the Zenodo page of the archive has the title, author and introduction consistent with the paper.

Please then let me know the DOI once it is done. Thanks!
