### 2025-05-22 SPARC-X/SPARC-X-API#99 @alchem0x2A
@jacksund I believe the new changes in upcoming v1.0.7 addresses your concerns and from suggestion of the editor I'm now closing this issue. Please feel free to re-open.
