### 2025-05-23 #7747 @alchem0x2A
Author checklist
        The latest version is [v1.0.7](https://github.com/SPARC-X/SPARC-X-API/tree/v1.0.7)
        The DOI is 10.5281/zenodo.15493021
