### 2025-05-27 #7747 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
