+++ paper/paper.bib
@@ -85,7 +85,7 @@
 	month = feb,
 	pages = 101295,
 	publisher = {Elsevier BV},
-	title = {Version 2.0.0 - M-SPARC: Matlab-Simulation Package for Ab-initio Real-space Calculations},
+	title = {Version 2.0.0 - {M}-{SPARC}: Matlab-Simulation Package for Ab-initio Real-space Calculations},
 	volume = 21,
 	year = 2023,
 	url = {http://dx.doi.org/10.1016/j.softx.2022.101295},
@@ -282,7 +282,7 @@
  	doi = {10.1038/s41592-019-0506-8},
  	pages = {670--673},
  	number = {8},
- 	journaltitle = {Nat Methods},
+ 	journaltitle = {Nature Methods},
  	author = {Bonomi, Massimiliano and Bussi, Giovanni and Camilloni, Carlo and Tribello, Gareth A. and Banáš, Pavel and Barducci, Alessandro and Bernetti, Mattia and Bolhuis, Peter G. and Bottaro, Sandro and Branduardi, Davide and Capelli, Riccardo and Carloni, Paolo and Ceriotti, Michele and Cesari, Andrea and Chen, Haochuan and Chen, Wei and Colizzi, Francesco and De, Sandip and De La Pierre, Marco and Donadio, Davide and Drobot, Viktor and Ensing, Bernd and Ferguson, Andrew L. and Filizola, Marta and Fraser, James S. and Fu, Haohao and Gasparotto, Piero and Gervasio, Francesco Luigi and Giberti, Federico and Gil-Ley, Alejandro and Giorgino, Toni and Heller, Gabriella T. and Hocky, Glen M. and Iannuzzi, Marcella and Invernizzi, Michele and Jelfs, Kim E. and Jussupow, Alexander and Kirilin, Evgeny and Laio, Alessandro and Limongelli, Vittorio and Lindorff-Larsen, Kresten and Löhr, Thomas and Marinelli, Fabrizio and Martin-Samos, Layla and Masetti, Matteo and Meyer, Ralf and Michaelides, Angelos and Molteni, Carla and Morishita, Tetsuya and Nava, Marco and Paissoni, Cristina and Papaleo, Elena and Parrinello, Michele and Pfaendtner, Jim and Piaggi, Pablo and Piccini, GiovanniMaria and Pietropaolo, Adriana and Pietrucci, Fabio and Pipolo, Silvio and Provasi, Davide and Quigley, David and Raiteri, Paolo and Raniolo, Stefano and Rydzewski, Jakub and Salvalaglio, Matteo and Sosso, Gabriele Cesare and Spiwok, Vojtěch and Šponer, Jiří and Swenson, David W. H. and Tiwary, Pratyush and Valsson, Omar and Vendruscolo, Michele and Voth, Gregory A. and White, Andrew and {The PLUMED consortium}},
  	urldate = {2024-09-18},
  	date = {2019-08},
@@ -310,7 +310,7 @@
 @article{enkovaara_gpaw_1_2011,
   author       = {Enkovaara, Jussi and Romero, Nichols A. and Shende,
                   Sameer and Mortensen, Jens J.},
-  title	       = {GPAW - massively parallel electronic structure
+  title	       = {{GPAW} - massively parallel electronic structure
                   calculations with Python-based software},
   journal      = {Procedia Computer Science},
   year	       = 2011,
@@ -456,7 +456,7 @@
                   Setten, Michiel J. and Van Troeye, Benoit and
                   Verstraete, Matthieu J. and Zérah, Gilles and
                   Zwanziger, Josef W.},
-  title	       = {The Abinit project: Impact, environment and recent
+  title	       = {The {A}binit project: Impact, environment and recent
                   developments},
   journal      = {Computer Physics Communications},
   year	       = 2020,
[…]
