+++ paper/paper.md
@@ -57,8 +57,8 @@ Matlab (M-SPARC [@xu_m-sparc-1.0_2020; @zhang_m-sparc-2.0_2023]) and C/C++
 of SPARC’s feature set has created the need for a fully functional
 interface to drive SPARC in high-throughput calculations.  Here we
 introduce SPARC-X-API, a Python package designed to bridge the SPARC-X
-project with broader computational frameworks. Built on the atomic
-simulation environment (ASE [@larsen_ase_2017]) standard, the SPARC-X-API
+project with broader computational frameworks. Built on the Atomic
+Simulation Environment (ASE [@larsen_ase_2017]) standard, the SPARC-X-API
 allows users to handle SPARC file formats and run SPARC calculations
 through the same interface as with other ASE-compatible DFT packages.
 Beyond standard ASE capabilities, SPARC-X-API provides additional
@@ -119,7 +119,7 @@ user-friendly and competitive with state-of-the-art plane-wave
 codes. The philosophy of the SPARC-X project is to provide codes that
 are highly efficient and portable (i.e., straightforward to install and
 use across various computational environments). The codes also seek to
-be user-friendly and developer-friendly to facilitate the
+be both user- and developer-friendly to facilitate the
 implementation of new algorithms. In line with this, SPARC-X offers
 real-space DFT algorithms through two implementations: 1) Matlab-based
 M-SPARC [@xu_m-sparc-1.0_2020; @zhang_m-sparc-2.0_2023] for algorithm
@@ -162,7 +162,7 @@ for example high-throughput dynamics simulations by i-PI
 active machine learning frameworks including FineTuna
 [@musielewicz_finetuna_2022], powered by state-of-art neural network
 interatomic potentials such as FAIR-Chem
-[https://github.com/FAIR-Chem/fairchem](https://github.com/FAIR-Chem/fairchem)
+([https://github.com/FAIR-Chem/fairchem](https://github.com/FAIR-Chem/fairchem))
 and MACE-MP [@ilyes_mace_2023] model series.  A summary of the role
 SPARC-X-API in the SPARC-X project is shown in
 \autoref{fig:sparc-overview}.  In addition to the capabilities
@@ -203,7 +203,7 @@ treating each calculation directory as a "SPARC bundle". The
 this bundle, ensuring that all necessary input and output files are
 properly handled. By default, the SPARC-X-API also copies relevant
 pseudopotential files into the calculation directory, making the SPARC
-bundle portable across different machines. From version 2.0 onwards,
+bundle portable across different machines. From version 1.0.7 onwards,
 the SPARC-X-API leverages the new features introduced in ASE version 3.23
 to register as an external I/O format, allowing reading and writing
 SPARC files directly using `ase.io` submodule:
@@ -347,7 +347,7 @@ computational setups.
 The design of the SPARC protocol allows insertion of bidirectional
 additional routines between two DFT calls, allowing further control
 over the low-level C/C++ code.
-Figure \autoref{fig:socket-hetero} summarizes the
+\autoref{fig:socket-hetero} summarizes the
 server-client setup across hybrid computing platforms.
 
 ![Example of socket communication across hybrid computing platforms using SPARC-X-API
