+++ paper/paper.bib
@@ -419,7 +419,7 @@
                   and Timrov, I and Thonhauser, T and Umari, P and
                   Vast, N and Wu, X and Baroni, S},
   title	       = {Advanced capabilities for materials modelling with
-                  Quantum ESPRESSO},
+                  {Q}uantum {ESPRESSO}},
   journal      = {Journal of Physics: Condensed Matter},
   year	       = 2017,
   volume       = 29,
@@ -602,7 +602,7 @@
 
 @article{bowler_order_n_dft_2012,
   author       = {Bowler, D R and Miyazaki, T},
-  title	       = {O(N) methods in electronic structure
+  title	       = {{O}({N}) methods in electronic structure
                   calculations},
   journal      = {Reports on Progress in Physics},
   year	       = 2012,
