### 2025-05-27 #7747 @alchem0x2A
@zhubonan We have fixed a few formatting typos in the main text and reference part. I would appreciate you use the following for the final version before accepting

- software github version: v1.0.8
- zenodo doi: 10.5281/zenodo.15529945
