### 2025-05-30 #7747 @alchem0x2A
@zhubonan Is there anything we should do before the manuscript moved to accepted status? One coauthor asks whether the doi for the publication can already be referenced. Thx!
