### 2025-06-01 #7747 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
