### 2025-06-04 #7747 @alchem0x2A
@zhubonan the proof looks good to me!
