### 2025-06-04 2448a15
bump setup.py version; allow parser to handle HUBBARD
