### 2025-06-04 07bf2d1
allow hubbard block in .ion file
