### 2025-06-04 c97d73b
clean up and add hubbard parameter from calc side
