### 2025-06-04 ea5179d
remove some print statements
