### 2025-06-05 16f498a
typo in hubbard block
