### 2025-06-06 2ddc15a
fix bug in hubbard flag writing in io code
