### 2025-06-06 75d7a8a
add test example and unittest for hubbard
