+++ tests/test_hubbard.py
@@ -1,0 +1,185 @@
+import os
+import tempfile
+from pathlib import Path
+
+import numpy as np
+import pytest
+
+curdir = Path(__file__).parent
+repo_dir = curdir.parent
+test_output_dir = curdir / "outputs"
+
+
+def test_parse_multiple_hubbard_atoms():
+    """_read_ion should return a data dict
+    containing multiple tuples of atomic U-correction
+    values
+    """
+    from sparc.sparc_parsers.ion import _read_ion, _write_ion
+
+    ion_content = """
+#=========================
+# format of ion file
+#=========================
+# ATOM_TYPE: <atom type name>
+# PSEUDO_POT: <path/to/pseudopotential/file>
+# N_TYPE_ATOM: <num of atoms of this type>
+# COORD:
+# <xcoord> <ycoord> <zcoord>
+# ...
+# RELAX:
+# <xrelax> <yrelax> <zrelax>
+# ...
+
+# Reminder: when changing number of atoms, change the RELAX flags accordingly
+#           as well.
+
+ATOM_TYPE: Ti                # atom type
+PSEUDO_POT: ../../../psps/22_Ti_12_2.0_2.8_pbe_n_v1.0.psp8
+N_TYPE_ATOM: 2               # number of atoms of this type
+COORD_FRAC:                  # coordinates follows
+0.000000 0.500000 0.250000
+0.500000 0.000000 0.750000
+
+ATOM_TYPE: Cr                # atom type
+PSEUDO_POT: ../../../psps/24_Cr_14_1.7_2.1_pbe_n_v1.0.psp8
+N_TYPE_ATOM: 2               # number of atoms of this type
+COORD_FRAC:                  # coordinates follows
+0.000000 0.000000 0.000000
+0.500000 0.500000 0.500000
+
+ATOM_TYPE: O               # atom type
+PSEUDO_POT: ../../../psps/08_O_6_1.2_1.4_pbe_n_v1.0.psp8
+N_TYPE_ATOM: 8               # number of atoms of this type
+COORD_FRAC:                  # coordinates follows
+0.000000 0.500000 0.460598
+0.500000 0.500000 0.297638
+0.000000 0.500000 0.039402
+0.000000 0.000000 0.202362
+0.500000 0.000000 0.960598
+0.000000 0.000000 0.797638
+0.500000 0.000000 0.539402
+0.500000 0.500000 0.702362
+
+HUBBARD:
+#U_ATOM_TYPE: Ti
+#U_VAL: 0 0 0.1 0 # U values in hartrees for s p d f orbitals (valence only)
+
+U_ATOM_TYPE: Cr
+U_VAL: 0 0 0.2 0 # U values in hartrees for s p d f orbitals (valence only)
+
+U_ATOM_TYPE: Ti
+U_VAL: 0 0 0.1 0 # U values in hartrees for s p d f orbitals (valence only)
+    """
+    with tempfile.TemporaryDirectory() as tempdir:
+        tempdir = Path(tempdir)
+        test_ion = tempdir / "test.ion"
+        test_ion_w = tempdir / "test_w.ion"
+        with open(test_ion, "w") as fd:
+            fd.write(ion_content)
+
+        data_dict = _read_ion(test_ion)
+        ion_dict = data_dict["ion"]
+        assert "extra" in ion_dict
+        assert "hubbard" in ion_dict["extra"]
+        hubbard_u_pairs = ion_dict["extra"]["hubbard"]
+        assert len(hubbard_u_pairs) == 2
+        # Atom types are read in sequence
+        # comments should be ommitted
+        assert hubbard_u_pairs[0]["U_ATOM_TYPE"] == "Cr"
[…]
