### 2025-06-06 232af3b
test hubbard writting with custom input
