### 2025-06-06 bb5e51c
add hubbard parameter checker
