### 2025-06-06 a7667b8
test for hubbard U without parameter
