### 2025-06-06 4cae0f8
force int conversion for HUBBARD_FLAG (CI)
