### 2025-06-06 837e3a1
add quicktest debugging
