### 2025-06-06 302de60
unset doc path in sparc test
