+++ .github/workflows/unit_test.yml
@@ -39,12 +39,13 @@ jobs:
         conda info
         flake8 sparc/ --count --select=E9,F63,F7,F82 --show-source --statistics
         flake8 sparc/ --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Run quicktest (debug for API version, setup etc)
-      run: |
-        python -m sparc.quicktest
     - name: Test with pytest (without all SPARC examples)
       run: |
         export ASE_SPARC_COMMAND="mpirun -n 1 sparc"
+        # Do not use default SPARC doc
+        unset SPARC_DOC_PATH
+        # Run a quicktest
+        python -m sparc.quicktest
         coverage run -a -m pytest -svv tests/
         coverage json --omit="tests/*.py"
         coverage html --omit="tests/*.py"
