### 2025-06-06 5c06e95
add raw result parser for Hubbard
