### 2025-06-07 bf2c3e1
get rid of pkg_resources and change to importlib
