### 2025-06-07 a732f61
manual disable api update CI
