+++ .github/workflows/update_api.yml
@@ -17,8 +17,10 @@ jobs:
       run:
         shell: bash -l {0}
     runs-on: ubuntu-latest
-    # Only run the create-PR action o upstream
-    # if: github.repository == "SPARC-X/SPARC-X-API"
+    if: |
+      # Only run the create-PR action o upstream for automatic workflows
+      # the fork maintainer can run using dispatch though
+      !github.event.repository.fork || github.event_name == 'workflow_dispatch'
 
     steps:
     - uses: actions/checkout@v4
@@ -52,7 +54,7 @@ jobs:
       continue-on-error: true
     - name: Create or Update Pull Request
       # Manually disable to make sure it can rerun
-      #if: env.UPDATE_NEEDED == 'true'
+      if: env.UPDATE_NEEDED == 'true'
       uses: peter-evans/create-pull-request@v7
       with:
         token: ${{ secrets.GITHUB_TOKEN }}
