### 2025-06-07 fafcf9a
Use a automated version check for setup.py
