### 2025-06-07 #7747 @alchem0x2A
@Kevin-Mattheus-Moerman Thank you!

Since the paper source code exists in the master branch of the repo, we uploaded a new version of the software for the archive
- Github version v1.1.0
- Zenodo archive 10.5281/zenodo.15616778

TODO lists

- [x] Please spell out USA in your affiliations as United States of America

- [x] Please edit the archive listed license tag to match v1.0.8 (i.e. include the v).
